Environmental Justice Scorecard
The Phase One Scorecard was released on April 21, 2023. It established a baseline of actions taken by federal agencies in 2021 and 2022 to help achieve the Biden-Harris Administration’s environmental justice goals. The Phase One Scorecard reported on the progress of 24 federal agencies in the following areas: advancing the President’s Justice40 Initiative; implementing and enforcing environmental and civil rights laws; and embedding environmental justice throughout the federal government.